refactor: extract jobs routes to routes/jobs.js
